+++ api/src/types.ts
@@ -31,3 +31,14 @@ export type TiddlConfig = {
     country_code: string;
   };
 };
+
+export type LogType = {
+  output: string;
+  output_history: string[];
+  link: string;
+  is_athenticated: boolean;
+  process?: ChildProcessWithoutNullStreams;
+  status?: "finished" | "error" | "auth" | undefined;
+  loading?: boolean;
+  error?: boolean;
+};