+++ app/src/index.tsx
@@ -5,7 +5,9 @@ import { ThemeProvider } from "@emotion/react";
 import { createTheme, CssBaseline } from "@mui/material";
 
 import PrivateRoute, { ROUTE_LOGIN } from "./components/Security/PrivateRoute";
+import { APIFetcherProvider } from "./provider/ApiFetcherProvider";
 import { AuthProvider } from "./provider/AuthProvider";
+import { ConfigProvider } from "./provider/ConfigProvider";
 import App from "./App";
 import Login from "./Login";
 
@@ -28,14 +30,18 @@ root.render(
     <ThemeProvider theme={darkTheme}>
       <CssBaseline />
       <BrowserRouter>
-        <AuthProvider>
-          <Routes>
-            <Route path={ROUTE_LOGIN} element={<Login />} />
-            <Route element={<PrivateRoute />}>
-              <Route path="/" element={<App />} />
-            </Route>
-          </Routes>
-        </AuthProvider>
+        <APIFetcherProvider>
+          <ConfigProvider>
+            <AuthProvider>
+              <Routes>
+                <Route path={ROUTE_LOGIN} element={<Login />} />
+                <Route element={<PrivateRoute />}>
+                  <Route path="/" element={<App />} />
+                </Route>
+              </Routes>
+            </AuthProvider>
+          </ConfigProvider>
+        </APIFetcherProvider>
       </BrowserRouter>
     </ThemeProvider>
   </React.StrictMode>,