+++ app/src/provider/AuthProvider.tsx
@@ -7,8 +7,9 @@ import React, {
 } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { ROUTE_LOGIN } from "src/components/Security/PrivateRoute";
-import { auth, is_auth_active } from "src/server/queryApi";
 import { ApiReturnType, AuthType, CheckAuthType } from "src/types";
+
+import { useApiFetcher } from "./ApiFetcherProvider";
 
 export const LOCALSTORAGE_TOKEN_KEY = "tidarr-token";
 
@@ -24,6 +25,9 @@ const AuthContext = React.createContext<AuthContextType>({} as AuthContextType);
 export function AuthProvider({ children }: { children: ReactNode }) {
   const [isAuthActive, setIsAuthActive] = useState<boolean>();
   const { pathname } = useLocation();
+  const {
+    actions: { auth, is_auth_active },
+  } = useApiFetcher();
   const navigate = useNavigate();
 
   const isAccessGranted = useMemo(
@@ -33,6 +37,7 @@ export function AuthProvider({ children }: { children: ReactNode }) {
 
   const check = async () => {
     const response = await is_auth_active();
+
     setIsAuthActive(response && (response as CheckAuthType).isAuthActive);
     if (pathname === ROUTE_LOGIN) navigate("/");
   };