+++ e2e/tests/search.spec.ts
@@ -68,7 +68,7 @@ test("Tidarr search : Should see albums results", async ({ page }) => {
 
   await expect(
     page.locator("#full-width-tabpanel-1 .MuiGrid-item").first(),
-  ).toHaveScreenshot({ maxDiffPixelRatio: 0.1 });
+  ).toHaveScreenshot();
 
   // Test pager
 
@@ -97,7 +97,7 @@ test("Tidarr search : Should see artists results", async ({ page }) => {
 
   await expect(
     page.locator("#full-width-tabpanel-2 .MuiGrid-item").first(),
-  ).toHaveScreenshot({ maxDiffPixelRatio: 0.1 });
+  ).toHaveScreenshot();
 
   // Test pager
 
@@ -155,7 +155,7 @@ test("Tidarr search : Should see tracks results", async ({ page }) => {
 
   await expect(
     page.locator("#full-width-tabpanel-3 .MuiGrid-item").first(),
-  ).toHaveScreenshot({ maxDiffPixels: 10 });
+  ).toHaveScreenshot();
 
   // Test pager
 
@@ -199,7 +199,8 @@ test("Tidarr search : Should see quality filtered results", async ({
 
   // Test localstorage persistence
 
-  await page.reload();
+  await runSearch("artist:3634161:The Beatles", page);
+
   await expect(
     await page.getByRole("button", { name: "Lossless" }),
   ).toHaveAttribute("aria-pressed", "true");